fix: aliases in namespace imports (#418) (#443)
* fix: aliases in namespace imports (#418)

Issue #418 appeared to boil down to missing support for _namespace
imports_ (i.e., those of the form `import * from Xyz in 'xyz'`).

This commit adds that support, along with a unit test and a change
in the `react-vite` example that relies on the new functionality.

* fix: linting issues

Fixing linting issues occurring during CI build.